Feature: added evaluation functions
An evaluation function is used to change how the error is calculated  when setting a ERROR_LIMIT criteria from the error calculated during backpropagation.

As an example, this can be used during classification to set the error equal to the categorical prediction accuracy rather  than the error calculcated by the cross entropy (which might not convey as much meaning).